+++ crates/astria-composer/src/composer.rs
@@ -25,13 +25,6 @@ use astria_eyre::eyre::{
 };
 use tokio::{
     net::TcpListener,
-    sync::{
-        mpsc::{
-            error::SendTimeoutError,
-            Sender,
-        },
-        watch,
-    },
     task::JoinError,
 };
 use tokio_util::task::JoinMap;
@@ -173,7 +166,8 @@ impl Composer {
     }
 
     /// Returns the socker address the grpc collector is served over
-    /// # Error: Returns an error if the listener is not bound
+    /// # Errors
+    /// Returns an error if the listener is not bound
     pub fn grpc_collector_local_addr(&self) -> io::Result<SocketAddr> {
         self.grpc_collector_listener.local_addr()
     }
@@ -327,47 +321,5 @@ fn report_exit(task_name: &str, outcome: Result<eyre::Result<()>, JoinError>) {
         Err(error) => {
             error!(%error, task = task_name, "task failed to complete");
         }
-    }
-}
-
-#[async_trait::async_trait]
-impl ComposerService for executor::Handle {
-    async fn submit_sequence_actions(
-        &self,
-        request: Request<SubmitSequenceActionsRequest>,
-    ) -> Result<Response<()>, tonic::Status> {
-        let submit_sequence_actions_request = request.into_inner();
-        if submit_sequence_actions_request.sequence_actions.is_empty() {
-            return Err(tonic::Status::invalid_argument(
-                "No sequence actions provided",
-            ));
-        }
-
-        // package the sequence actions into a SequenceAction and send it to the searcher
-        for sequence_action in submit_sequence_actions_request.sequence_actions {
-            let sequence_action = SequenceAction {
-                rollup_id: RollupId::from_unhashed_bytes(sequence_action.rollup_id),
-                data: sequence_action.tx_bytes,
-                fee_asset_id: default_native_asset_id(),
-            };
-
-            match self
-                .get()
-                .send_timeout(sequence_action, Duration::from_millis(500))
-                .await
-            {
-                Ok(()) => {}
-                Err(SendTimeoutError::Timeout(_seq_action)) => {
-                    return Err(tonic::Status::deadline_exceeded(
-                        "timeout while sending txs to searcher",
-                    ));
-                }
-                Err(SendTimeoutError::Closed(_seq_action)) => {
-                    return Err(tonic::Status::unavailable("searcher is not available"));
-                }
-            }
-        }
-
-        Ok(Response::new(()))
     }
 }
